requirements.txt
google-genai
python-dotenv
mcp[cli]
pydantic
starlette
uvicorn
httpx
